pyproject.toml
[project]
name = "gcp-tutorial-aims-gemma-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "datasets>=4.5.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "setuptools==81.0",
    "tensorflow>=2.20.0",
    "tensorflow-tensorboard>=1.5.1",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "wandb>=0.25.0",
]
